+++ hubdle-settings-gradle-plugin/main/kotlin/hubdleSettingsPlugin.accessor.kt
@@ -1,0 +1,10 @@
+import com.javiersc.hubdle.settings.HubdleSettingsDslMarker
+import com.javiersc.hubdle.settings.HubdleSettingsExtension
+import com.javiersc.hubdle.settings.hubdleSettings
+import org.gradle.api.Action
+import org.gradle.api.initialization.Settings
+
+@HubdleSettingsDslMarker
+public fun Settings.hubdleSettings(action: Action<HubdleSettingsExtension> = Action {}) {
+    action.execute(hubdleSettings)
+}